Fix converter

+++ src/main/java/com/ruralhousejsf/converter/RuralHouseConverter.java
@@ -13,7 +13,7 @@ public class RuralHouseConverter extends SelectItemsConverter  {
 
 	@Override
 	public String getAsString(FacesContext context, UIComponent component, Object value) {		
-		Integer id = (value instanceof RuralHouse) ? ((RuralHouse) value).getId() : null;
+		Long id = (value instanceof RuralHouse) ? ((RuralHouse) value).getId() : null;
 		return (id != null) ? String.valueOf(id) : null;
 	}
